fix: move search output logic to a new component
